Moved ElementParameterGet and ElementParameterGet to ElementParameter.cs
